upgrade exercise

+++ Capitolo-3/Testing-1/src/Forma.java
@@ -5,6 +5,6 @@ enum TipoForma{
 
 public class Forma {
 
-    public double calcolaArea(){ return 0;}
+    public Double calcolaArea(){ return (double) 0;}
 
 }
